+++ controller/konnect/ops/zz_generated_ops_delete.go
@@ -25,6 +25,8 @@ func DeleteGeneratedOps[
 	e TEnt,
 ) error {
 	switch ent := any(e).(type) {
+	case *konnectv1alpha1.AIGatewayAgent:
+		return deleteAIGatewayAgent(ctx, sdk.GetAIGatewayAgentsSDK(), ent)
 	case *konnectv1alpha1.AIGatewayControlPlane:
 		return deleteAIGatewayControlPlane(ctx, sdk.GetAIGatewaysSDK(), ent)
 	case *configurationv1alpha1.AIGatewayDataPlaneCertificate: